@@ -1,4 +1,4 @@
-package net.oleksin.socket.commandfactory.command;
+package net.oleksin.socket.command;
 
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
@@ -6,42 +6,44 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.util.Collections;
+import java.util.Set;
+
+import net.oleksin.Context;
 
 public class LsCommand implements Command {
   @Override
-  public void execute(final CommandContext commandContext) {
-    if (commandContext.isPathNull()) {
-      printToOut(commandContext);
+  public void execute(final Context context) {
+    if (context.isPathNull()) {
+      printException(context);
     } else {
-      readFromFile(commandContext);
+      walkFileTree(context);
     }
   }
   
-  private void readFromFile(CommandContext commandContext) {
+  private void walkFileTree(Context context) {
     try {
-      Files.walkFileTree(commandContext.getPath(),
-              Collections.emptySet(), 1, new MyFileVisitor(commandContext));
+      Files.walkFileTree(context.getPath(),
+              Set.of(), 1, new MyFileVisitor(context));
     } catch (IOException e) {
-      printToOut(commandContext);
+      printException(context);
     }
   }
   
-  private void printToOut(CommandContext commandContext) {
-    commandContext.addToOutput(
+  private void printException(Context context) {
+    context.printLn(
             "Directory does not exist! Please, input directory!");
   }
 
   private class MyFileVisitor extends SimpleFileVisitor<Path> {
-    private final CommandContext commandContext;
+    private final Context context;
 
-    public MyFileVisitor(CommandContext commandContext) {
-      this.commandContext = commandContext;
+    public MyFileVisitor(Context context) {
+      this.context = context;
     }
 
     @Override
     public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) {
-      commandContext.addToOutput(
+      context.printLn(
               String.format("%-25.25s:\t%tc", file.getFileName(),
                       attrs.lastModifiedTime().toMillis()));
       return FileVisitResult.CONTINUE;
